diplom/requirements.txt
django
djangorestframework
geopy
pillow
psycopg2